Move files for better modularization
