+++ vtuber/abstract_random_test.cpp
